+++ src/player.cpp
@@ -1114,6 +1114,97 @@ void Player::resetStreamTime()
     soloud.resetClockedAnchor();
 }
 
+double Player::getEngineTime()
+{
+    return soloud.getEngineTime();
+}
+
+PlayerErrors Player::playScheduled(
+    unsigned int soundHash,
+    unsigned int &handle,
+    double atTime,
+    double duration,
+    unsigned int busId,
+    float volume,
+    float pan)
+{
+    ActiveSound *sound = findByHash(soundHash);
+
+    if (sound == nullptr)
+        return soundHashNotFound;
+
+    // A BufferStream using `release` buffer type can only have one instance.
+    if (sound->soundType == SoundType::TYPE_BUFFER_STREAM &&
+        static_cast<SoLoud::BufferStream *>(sound->sound.get())->getBufferingType() == BufferingType::RELEASED &&
+        sound->handle.size() > 0)
+    {
+        return bufferStreamCanBePlayedOnlyOnce;
+    }
+
+    // Check if playing this sound will exceed the maximum number of voice counts. If true, then
+    // check if [soudHash] has other instances playing. If true remove the first and play the new one.
+    // If no other instances are playing, this sound cannot be played and return an error.
+    // Issue https://github.com/alnitak/flutter_soloud/issues/204
+    if (getActiveVoiceCount_internal() >= getMaxActiveVoiceCount())
+    {
+        if (sound->handle.size() > 0)
+        {
+            stop(sound->handle[0].handle);
+        }
+        else
+        {
+            return PlayerErrors::maxActiveVoiceCountReached;
+        }
+    }
+
+    // Ensure miniaudio device is started if it's stopped, ie by an interruption.
+    soloud.resume();
+
+    handle = 0;
+    SoLoud::handle newHandle = 0;
+    if (busId == 0)
+    {
+        newHandle = soloud.playScheduled(
+            atTime, *sound->sound.get(), volume, pan, 0);
+    }
+    else
+    {
+        auto it = busMap.find(busId);
+        if (it != busMap.end())
+            newHandle = it->second.bus.playScheduled(
+                atTime, *sound->sound.get(), volume, pan);
+        else
+            return PlayerErrors::busIdNotFound;
+    }
+
+    if (newHandle != 0)
+    {
+        sound->handle.push_back({newHandle, MAX_DOUBLE, false});
+        if (duration > 0.0)
+        {
+            soloud.scheduleStopAt(newHandle, atTime + duration);
+        }
+        // Check if this buffer has enough data to be played
+        if (sound->soundType == SoundType::TYPE_BUFFER_STREAM)
+        {
+            static_cast<SoLoud::BufferStream *>(sound->sound.get())->checkBuffering(0);
+        }
+    }
+    handle = newHandle;
+    return PlayerErrors::noError;
+}
+
+void Player::stopScheduled(unsigned int handle, double atTime)
+{
+    soloud.scheduleStopAt(handle, atTime);
+}
+
+void Player::fadeScheduled(unsigned int handle, double atTime, float to,
+                           double fadeTime, bool thenStop)
+{
+    soloud.scheduleFadeAt(handle, atTime, to, fadeTime, thenStop);
+}
+
 void Player::stop(unsigned int handle)
 {
     soloud.stop(handle);